NAMESPACE: Drop extraneous verbs.
Based on a survey of the dscr repos in production, prune the unused
verbs from the codebase, making a few modifications for logical
consistency. This drops the NAMESPACE exports from 47 to 18.

When using the internal utilities directly as a part of dscr
development, these dropped verbs can be loaded with
`devtools::load_all()` or Ctrl+Shift+L in RStudio.