seeded_function_call: Refactor run_scenario and run_method.
Refactor `run_scenario` and `run_method` to replace their loading of
data with readRDS, seed-setting, and try-wrapped function calls with a
call to `seeded_function_call` instead. The game plan is to develop an
execution plan that can tabulate the `seeded_function_call` calls for a
given dscr phase in advance. Once we can develop and test such a data
structure, it could be passed to a `batchMap` call for parallel
processing.